+++ bad_test_os.yaml
@@ -1,13 +1,13 @@
 apiVersion: v1
 kind: Pod
 metadata:
-  name: super_pod
+  name: ""
   namespace: super_service
   labels:
     dc: us-west-1
     group: gamma
 spec:
-  os: qwerty
+  os: pldQ
   containers:
     - name: my_container_name
       image: registry.bigbrother.io/baseimage:v1.2.0
@@ -27,5 +27,5 @@ spec:
           cpu: 2
           memory: "1Gi"
         requests:
-          cpu: 1
+          cpu: "1"
           memory: "500Mi"
